adds alternative method to downloading chain data
For some reason the 'URL=...' was giving me 'parse error: Invalid numeric literal at line 1, column 7' so i had to find an alternative method. Would be nice to include alternative methods in the docs, like i propose. 

NOTE: Placement of the text i added is definitely not the best, so please adjust that.